% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_IEA.R
\name{calculate_IEA}
\alias{calculate_IEA}
\title{Applies IEA to get a baseline no-synergy/antagonism mixture DER.}
\usage{
calculate_IEA(dose, LET, ratios, model = "NTE", coef = list(NTE =
  HZE_nte_model_coef, TE = HZE_te_model_coef, lowLET = low_LET_model_coef),
  ders = list(NTE = calibrated_HZE_nte_der, TE = calibrated_HZE_te_der,
  lowLET = calibrated_low_LET_der), calculate_dI = c(NTE =
  .calculate_dI_nte, TE = .calculate_dI_te), phi = 2000)
}
\arguments{
\item{dose}{Numeric vector corresponding to the sum dose in cGy.}

\item{LET}{Numeric vector of all LET values, must be length n.}

\item{ratios}{Numeric vector of all dose ratios, must be length n.}

\item{model}{String value corresponding to the model to be used, either "NTE" or "TE".}

\item{coef}{Named list of numeric vectors containing coefficients for one-ion DERs.}

\item{ders}{Named list of functions containing relevant one-ion DER models.}

\item{calculate_dI}{Named vector of functions to calculate dI depending on
the selected model.}

\item{phi}{Numeric value, used in NTE models.}
}
\value{
Numeric vector representing the estimated Harderian Gland
        prevalence from an IEA mixture DER constructed from the given
        one-ion DERs parameters.
}
\description{

}
\details{
Corresponding elements of ratios, LET should be associated with the
         same DER.
}
\examples{
calculate_IEA(.01 * 0:40, c(70, 195), c(1/2, 1/2))
calculate_IEA(.01 * 0:70, c(.4, 195), c(4/7, 3/7), model = "TE")

}
\author{
Dae Woong Ham, Edward Greg Huang <eghuang@berkeley.edu>
}
