% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_SEA.R
\name{calculate_SEA}
\alias{calculate_SEA}
\title{Applies Simple Effect Additivity to get a baseline mixture DER.}
\usage{
calculate_SEA(dose, LET, ratios, lowLET = FALSE, n = NULL)
}
\arguments{
\item{dose}{Numeric vector corresponding to the sum dose in cGy.}

\item{LET}{Numeric vector of all LET values, must be length n.}

\item{ratios}{Numeric vector of all dose ratios, must be length n.}

\item{lowLET}{Boolean of whether a low LET DER should be included in the mixture DER.}

\item{n}{Number of DERs, optional argument used to check parameter validity.}
}
\value{
Numeric vector representing the estimated Harderian Gland
        prevalence from a SEA mixture DER constructed from the given DER
        parameters.
}
\description{

}
\details{
Corresponding elements of ratios, LET should be associated with the
         same DER.
}
\examples{
calculate_SEA(.01 * 0:40, c(70, 195), c(1/2, 1/2), n = 2)
calculate_SEA(.01 * 0:70, c(0.4, 195), c(4/7, 3/7)) # RKS please explain n=NULL and lowLET = false a bit more. Also, what happens if we add more models in the cross validation?
}
