% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_der.R
\name{check_der}
\alias{check_der}
\title{Checks whether a DER satisfies the desired properties}
\usage{
check_der(DER, upper = 10^3, upperLET = 500, coeff = NULL, ...)
}
\arguments{
\item{DER}{Dose effect relationship function taking dose and LET as
arguments.}
}
\value{
Boolean.
}
\description{
Checks whether the input DER is nonnegative when dose is
             nonnegative.
}
\details{
Only tests LET values in [1, 500].
}
\examples{
ion_data <- load_ion_data("one_ion.csv")
HZE_nte_der <- make_der(ion_data, TRUE, TRUE)
check_der(HZE_nte_der)

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
