% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val.R
\name{cross_val}
\alias{cross_val}
\title{Applies cross validation to a mixture of ions with respect to
       a synergy theory.}
\usage{
cross_val(ions, model, prev, w)
}
\arguments{
\item{ions}{List of dataframes corresponding to ion.}

\item{model}{String, "NTE" or "TE" for non-targeted or targeted effects.}

\item{prev}{Numeric vector of observed prevalence values.}

\item{w}{Numeric vector of experimental weights.}
}
\value{
Numeric vector representing the estimated Harderian Gland
        prevalence from a SEA mixture DER constructed from the given DER
        parameters. RKS 5/17/2019. Again: why SEA?
}
\description{

}
\details{
Weight vector elements should correspond to dataframe element order.
         i.e. w[n] = ions[length(ions[:, 1]) / n][length(ions[:, 1]) % n]

Tested for the two examples below. General correctness is
         unverified and should be treated at some future data.
}
\examples{
NTE_cv <- cross_val(set_list,"NTE", HZE_data$Prev, HZE_data$NWeight)
TE_cv <- cross_val(set_list, "TE", HZE_data$Prev, HZE_data$NWeight)

}
