% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_monte_carlo.R
\name{.generate_ci}
\alias{.generate_ci}
\title{Generates confidence intervals for mixture baeline DER samples
             at a dose.}
\usage{
.generate_ci(n, dose_index, sample_curves, interval_length)
}
\arguments{
\item{n}{Numeric integer of the number of samples to be drawn.}

\item{dose_index}{Numeric integer of doses}

\item{sample_curves}{Numeric list of sampled mixture DER values.}

\item{interval_length}{Numeric double of the confidence interval width.}
}
\value{
Numeric length-two vector of an upper and lower bound for the
        confidence interval of a dose.
}
\description{

}
