% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea.R
\name{iea}
\alias{iea}
\title{Applies IEA to get a baseline no-synergy/antagonism mixture DER.}
\usage{
iea(dose, LET, ratios, E, dE = NULL, coeff = NULL, check = FALSE,
  ...)
}
\arguments{
\item{dose}{Numeric vector corresponding to the sum dose in cGy.}

\item{LET}{Numeric vector of all LET values, must be length n.}

\item{ratios}{Numeric vector of all dose ratios, must be length n.}

\item{E}{Vector of dose effect relationship functions.}

\item{dE}{Optional vector of functions corresponding to the derivatives of E.}

\item{check}{Optional boolean whether to check DERs in each run.}

\item{...}{Optional arguments to DER functions in E.}
}
\value{
Numeric vector representing the estimated Harderian Gland tumor
        prevalence from an IEA mixture DER constructed from the given
        one-ion DERs parameters.
}
\description{

}
\details{
Corresponding elements of ratios, LET should be associated with the
         same DER.
}
\examples{
ion_data <- load_ion_data("one_ion.csv")
HZE_nte_der <- make_der(ion_data, TRUE, TRUE)
check_der(HZE_nte_der)
# Iron and silicon two-ion mixture.
iea(0:100, c(70, 30), c(1/3, 2/3), rep(c(HZE_nte_der), 2))

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
