% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{loocv}
\alias{loocv}
\title{Applies leave-one-out cross validation to a mixture of ions with
       respect to the NTE or TE models used from Huang 2019.}
\usage{
loocv(ions, prev, wgts, model = c("NTE", "TE"), y_0 = 0.046404,
  phi = 2000)
}
\arguments{
\item{ions}{List of dataframes corresponding to ion.}

\item{prev}{Numeric vector of observed prevalence values.}

\item{model}{String, "NTE" or "TE" for non-targeted or targeted effects.}

\item{w}{Numeric vector of experimental weights.}
}
\value{
Numeric weighted mean errors.
}
\description{

}
\details{
Weight vector elements should correspond to dataframe element order.
         i.e. w[n] = ions[length(ions[:, 1]) / n][length(ions[:, 1]) % n]

Tested for the two examples below.
}
\examples{
##=================== Cross validation ====================#
# Seperate Data into 8 blocks, i.e. test/training sets:
HZE_data <- select(filter(ion_data, Z > 3), 1:length(ion_data))
data_len <- 1:length(HZE_data)
O_350 <- select(filter(HZE_data, LET == 20), data_len)
Ne_670 <- select(filter(HZE_data, LET == 25), data_len)
Si_260 <- select(filter(HZE_data, LET == 70), data_len)
Ti_1000 <- select(filter(HZE_data, LET == 100), data_len)
Fe_600 <- select(filter(HZE_data, LET == 193), data_len)
Fe_350 <- select(filter(HZE_data, LET == 253), data_len)
Nb_600 <- select(filter(HZE_data, LET == 464), data_len)
La_593 <- select(filter(HZE_data, LET == 953), data_len)
set_list <- list(O_350, Ne_670, Si_260, Ti_1000,
                 Fe_600, Fe_350, Nb_600, La_593)
actual_prev <- HZE_data$Prev

NTE_cv <- loocv(set_list,"NTE", HZE_data$Prev, HZE_data$NWeight)
TE_cv <- loocv(set_list, "TE", HZE_data$Prev, HZE_data$NWeight)
cv_table <- cbind(NTE_cv, TE_cv)
cv_table

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
