% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DER.R
\name{make_DER}
\alias{make_DER}
\title{Makes a dose effect relationship function for a specified dataset.}
\usage{
make_DER(data, HZE, NTE, phi = 2000, y_0 = 0.04604)
}
\arguments{
\item{data}{Data.frame of single ion data.}

\item{HZE}{Boolean for presence of HZE ions}

\item{NTE}{Boolean for presence of non-targeted effects}

\item{phi}{Numeric constant}

\item{y_0}{Numeric constant corresponding to background effect}
}
\value{
Function
}
\description{
Outputs a dose effect relationship (DER) for a single ion
             with parameters dose and LET. The ion may be a low-LET or HZE
             ion and non-targeted effects may be specified.
}
\details{
data must have a numeric column "dose" and a numeric weight column
         "NWeight".
}
\examples{

}
