% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model.R
\name{make_model}
\alias{make_model}
\title{Makes a dose effect relationship model for a specified dataset.}
\usage{
make_model(data, HZE, NTE = TRUE, phi = 2000, y_0 = 0.046404)
}
\arguments{
\item{data}{Data.frame of single ion data.}

\item{HZE}{Boolean for presence of HZE ions}

\item{NTE}{Boolean for presence of non-targeted effects}

\item{phi}{Numeric constant}

\item{y_0}{Numeric constant corresponding to background effect}
}
\value{
nls model
}
\description{
Outputs a dose effect relationship (DER) model for a single ion
             with parameters dose and LET. The ion may be a low-LET or HZE
             ion and non-targeted effects may be specified.
}
\details{
data must have a numeric column "dose" and a numeric weight column
         "NWeight".
}
\examples{
ion_data <- load_ion_data("one_ion.csv")

LLET_model    <- make_model(ion_data, FALSE)
HZE_nte_model <- make_model(ion_data, TRUE, TRUE)
HZE_te_model  <- make_model(ion_data, TRUE, FALSE)

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
