% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ions.R
\name{plot_ions}
\alias{plot_ions}
\title{Plots a mixed field.}
\usage{
plot_ions(dose, LET, ratios, E, models, dE = NULL, n = 20, ...)
}
\arguments{
\item{dose}{Numeric vector corresponding to the sum dose in cGy.}

\item{LET}{Numeric vector of all LET values, must be length n.}

\item{ratios}{Numeric vector of all dose ratios, must be length n.}

\item{E}{Vector of DER functions taking dose and LET as arguments.}

\item{...}{Optional arguments to DER functions in E.}
}
\value{

}
\description{

}
\details{

}
\examples{
ion_data      <- load_ion_data("one_ion.csv")
HZE_nte_der   <- make_der(ion_data, TRUE, TRUE)
HZE_nte_model <- make_model(ion_data, TRUE, TRUE)
plot_ions(0:100, c(193, 70), c(1/3, 2/3), E = c(HZE_nte_der, HZE_nte_der),
          models = list(HZE_nte_model, HZE_nte_model))

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
