% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_monte_carlo.R
\name{run_monte_carlo}
\alias{run_monte_carlo}
\title{Runs the Monte Carlo method on a baseline no-synergy/antagonism
       mixture DER.}
\usage{
run_monte_carlo(n = 200, dose, LET, ratios, model = "NTE",
  vcov = TRUE, interval_length = 0.95, seed = 100)
}
\arguments{
\item{n}{Numeric integer of the number of samples to be drawn.}

\item{dose}{Numeric vector of all total dose values to be evaluated.}

\item{LET}{Numeric vector of all LET values, must be length n.}

\item{ratios}{Numeric vector of dose proportions applied on component DERs.}

\item{model}{String value corresponding to the model to be used.}

\item{vcov}{Boolean for assessing inter-parameter correlation.}

\item{interval_length}{Numeric double of the confidence interval width.}

\item{seed}{Numeric value for pseudorandom generators.}
}
\value{
Named list representing lower and upper bounds for a Monte Carlo
        confidence interval for a mixture DER over an interval of doses.
}
\description{

}
\details{
Corresponding elements of ratios, LET should be associated with the
         same DER.
}
\examples{
ratios <- c(1/2, 1/2)
LET_vals <- c(195, 70)
run_monte_carlo(0:100, LET_vals, ratios)

}
\author{
Yimin Lin, Edward Greg Huang <eghuang@berkeley.edu>
}
