% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea.R
\name{sea}
\alias{sea}
\title{Applies Simple Effect Additivity to get a baseline mixture DER.}
\usage{
sea(dose, LET, ratios, E, coeff, ...)
}
\arguments{
\item{dose}{Numeric vector corresponding to the sum dose in cGy.}

\item{LET}{Numeric vector of all LET values, must be length n.}

\item{ratios}{Numeric vector of all dose ratios, must be length n.}

\item{E}{Vector of DER functions taking dose and LET as arguments.}

\item{...}{Optional arguments to DER functions in E.}
}
\value{
Numeric vector representing the estimated Harderian Gland
        prevalence from a SEA mixture DER constructed from the given DER
        parameters.
}
\description{

}
\details{
Corresponding elements of ratios, LET should be associated with the
         same DER.
}
\examples{
sea(.01 * 0:40, c(70, 195), c(1/2, 1/2), n = 2)
sea(.01 * 0:70, c(0.4, 195), c(4/7, 3/7))

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
