% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DER.R
\name{check_DER}
\alias{check_DER}
\title{Checks whether a DER satisfies the desired properties}
\usage{
check_DER(DER)
}
\arguments{
\item{DER}{Dose effect relationship function taking dose and LET as
arguments.}
}
\value{
Boolean.
}
\description{
Checks whether the input DER is nonnegative when dose is
             nonnegative.
}
\details{
Only tests LET values in [1, 500].
}
\examples{

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
