% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_monte_carlo.R
\name{.generate_samples}
\alias{.generate_samples}
\title{Generates mixture no-synergy/antagonism DER samples with
       parameters drawn from a Gaussian distribution.}
\usage{
.generate_samples(n, dose, LET, ratios, model, vcov,
  nte_model = HZE_nte_model, te_model = HZE_te_model,
  low_model = low_LET_model)
}
\arguments{
\item{n}{Numeric integer of the number of samples to be drawn.}

\item{dose}{Numeric vector of all total dose values to be evaluated.}

\item{LET}{Numeric vector of all LET values, must be length n.}

\item{ratios}{Numeric vector of dose proportions applied on component DERs.}

\item{model}{String value corresponding to the model to be used.}

\item{vcov}{Boolean for assessing inter-parameter correlation.}

\item{nte_model}{The input HZE nontargeted and targeted effects model.}

\item{te_model}{The input HZE targeted effects only model.}

\item{low_model}{The input low LET model.}
}
\value{
Numeric vector of sample mixture baseline DERs evaluated at the given doses.
}
\description{

}
