% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ion_data.R
\name{load_ion_data}
\alias{load_ion_data}
\title{Loads CSV dataset as a data.frame, dropping zero entries.}
\usage{
load_ion_data(filename)
}
\arguments{
\item{filename}{String of the CSV filename.}
}
\value{
Data.frame of the selected data file.
}
\description{
This is meant to handle single-ion CSV datasets.
}
\details{
The filename includes the path from the current working directory.
         CSV should have a numeric "dose" column.
}
\examples{
# Data used is that in 16Chang plus later NSRL data scored before 3/31/2019.
ion_data <- load_ion_data("one_ion.csv")
mix_data <- load_ion_data(read.csv("mix_ion.csv"))
# The two .csv files contain all input HG data except Y_0

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
