% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ion_data.R
\name{load_ion_data}
\alias{load_ion_data}
\title{Loads CSV dataset as a data.frame, dropping entries with zero.}
\usage{
load_ion_data(filename)
}
\arguments{
\item{filename}{String of the CSV filename.}
}
\value{
Data.frame of the selected data file.
}
\description{
Meant to handle single-ion CSV datasets.
}
\details{
The filename includes the path from the current working directory.
         CSV should have a numeric "dose" column. Numeric olumns with "Z"
         for atomic number and "NWeights" are required for compatibility
         with built in modeling functions.
}
\examples{
# Data used is that in 16Chang plus later NSRL data scored before 3/31/2019.
# The two .csv files contain all input HG data except Y_0
ion_data <- load_ion_data("one_ion.csv")
mix_data <- load_ion_data("mix_ion.csv")

}
\author{
Edward Greg Huang <eghuang@berkeley.edu>
}
